extracts SLO client
